backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Elevator Advocacy Platform Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django~=6.0",
    "djangorestframework",
    "requests",
    "python-dotenv",
    "django-cors-headers>=4.9.0",
    "google-genai>=1.72.0",
    "instructor>=1.15.1",
    "serpapi>=1.0.2",
    "types-requests>=2.31.0",
    "gunicorn>=25.3.0",
    "uvicorn>=0.44.0",
    "whitenoise>=6.12.0",
    "psycopg[binary]>=3.3.3",
    "dj-database-url>=3.1.2",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py"]
addopts = ["--tb=short", "--no-header", "-q"]

[tool.ruff]
line-length = 88
output-format = "concise"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
ignore_missing_imports = true
pretty = true

[tool.django-stubs]
django_settings_module = "config.settings"

[dependency-groups]
dev = [
    "django-stubs>=6.0.2",
    "djangorestframework-stubs>=3.16.9",
    "mypy>=1.20.1",
    "pytest>=9.0.3",
    "pytest-django>=4.12.0",
    "ruff>=0.15.10",
]
